renamed dev-server to dev
